requirements.txt
absl-py==2.3.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
deepdiff==7.0.1
distlib==0.3.9
Farama-Notifications==0.0.4
fcd-torch==1.0.7
fonttools==4.58.5
grpcio==1.73.1
gymnasium==0.28.1
h5py==3.14.0
jax-jumpy==1.0.0
kiwisolver==1.4.8
lightning==2.5.2
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
networkx==3.5
numba==0.61.2
ordered-set==4.1.0
overrides==7.7.0
pandas==2.3.1
pettingzoo==1.24.3
pillow==11.3.0
protobuf==6.31.1
pyparsing==3.2.3
pytz==2025.2
rdkit==2025.3.3
reportlab==4.4.4
rich==14.0.0
scipy==1.16.1
selfies==2.2.0
sensai-utils==1.4.0
setuptools==78.1.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tianshou==1.2.0
triton==3.1.0
tzdata==2025.2
virtualenv==20.31.2
Werkzeug==3.1.3
wheel==0.45.1
